Delete unused pages
