+++ client/src/layouts/Home.jsx
@@ -1,0 +1,5 @@
+export default function Home() {
+  return (
+    <div>Home</div>
+  )
+}
