--- client/src/pages/Home.jsx
@@ -1,5 +1,0 @@
-export default function Home() {
-  return (
-    <div>Home</div>
-  )
-}
